third_party/blink/web_tests/platform/win/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/neg.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] neg int32 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test neg int32 actual -2147483648 should be close enough to expected -2147483646 by ULP distance: expected a number less than or equal to 0 but got 2
Harness: the test ran to completion.
